test: create jestSetup file to center mocks
