+++ package.json
@@ -60,7 +60,9 @@
     "moduleDirectories": [
       "node_modules",
       "./src/test"
-    ]
+    ],
+    "modulePathIgnorePatterns": [".*/mockedData/.*"],
+    "setupFiles": ["<rootDir>/src/test/jestSetup.tsx"]
   },
   "private": true
 }
